requirements/_base.txt
Django>=5.0
graphene>=3.3
graphene-django>=3.2.2
graphene-django-optimizer>=0.10.0
django-model-utils>=4.3.1
djangorestframework>=3.15.1
python-decouple>=3.8
psycopg2>=2.9.6
django-log-request-id>=2.1.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
ipython>=8.18.1
setuptools>=69.0.2
django-permissions-policy>=4.18.0
django-environ>=0.11.2
drf-spectacular>=0.27.2
pillow>=10.4.0
factory-boy>=3.3.0
Faker>=26.0.0